[+] object and function first iteration
