+++ CMakeLists.txt
@@ -7,7 +7,7 @@ add_compile_definitions(
 
 set(CMAKE_CXX_STANDARD 23)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra -pedantic -Wno-unused-parameter")
 include_directories(${PROJECT_SOURCE_DIR}/includes)
 add_subdirectory(submodules/Lexer)
