+++ lib/types/objects/JsArray.cpp
@@ -13,18 +13,21 @@ namespace JS {
             JS::Function([self = shared_from_this()] (const std::vector<JS::Any>& args) {
 //                auto last = self->elements->back();
                 self->elements->pop_back();
-                std::cout << "lambda ope " << &self->properties << std::endl;
-                auto& lengthProperty = self->operator[]("length");
-                lengthProperty = JS::Any(static_cast<int>(self->elements->size()));
+                self->operator[]("length") = self->operator[]("length") - JS::Any(1);
                 return JS::Any(0);
             })));
+    }
+
+    JS::Any& Array::operator[](const std::string& key) {
+        return properties->operator[](key);
     }
 
     JS::Any& Array::operator[](size_t index) {
         if (index >= elements->size()) {
             elements->resize(index + 1, JS::Any(JS::Undefined{}));
+            this->operator[]("length") = this->operator[]("length") + JS::Any(1);
         }
-        properties->operator[]("length") = JS::Any(static_cast<int>(elements->size()));
+//        properties->operator[]("length") = JS::Any(static_cast<int>(elements->size()));
         return elements->operator[](index);
     }
 
@@ -42,34 +45,10 @@ namespace JS {
 
     void Array::push(const JS::Any& value) {
         elements->push_back(value);
-        properties->operator[]("length") = JS::Any(static_cast<int>(elements->size()));
+//        properties->operator[]("length") = JS::Any(static_cast<int>(elements->size()));
     }
 
     bool Array::isCallable() const {
         return false;
     }
 }
-
-
-//JS::Any& Array::operator[](size_t index) {
-//    if (index >= elements.size()) {
-//        elements.resize(index + 1, JS::Any(JS::Undefined{}));
-//    }
-//    properties["length"] = JS::Any(static_cast<int>(elements.size()));
-//    return elements[index];
-//}
-//
-//const JS::Any& Array::operator[](size_t index) const {
-//    if (index < elements.size()) {
-//        return elements[index];
-//    }
-//    throw std::out_of_range("Array index out of bounds.");
-//}
-//
-//JS::Any& Array::operator[](int index) {
-//    return this->operator[](static_cast<size_t>(index));
-//}
-//
-//const JS::Any& Array::operator[](int index) const {
-//    return this->operator[](static_cast<size_t>(index));
-//}
